Fixed strength again

+++ src/application/model/AI.java
@@ -4,8 +4,8 @@ import application.model.Board.Type;
 
 public class AI {
 
-	public int evaluateBoard(Board board) {
-		int strength = 0;
+	public double evaluateBoard(Board board) {
+		double strength = 0;
 		Piece[][] boardM = board.getBoard();
 		for(int i = 0; i < 8; i++)
 			for(int j = 0; j < 8; j++) {
